third_party/WebKit/LayoutTests/external/wpt/xhr/abort-event-order-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: The abort() method: abort and loadend events assert_equals: expected "upload.abort(0,0,false)" but got "upload.progress(0,0,false)"
Harness: the test ran to completion.
